introduce Pop type + lift pop scale to support billion-scale civs
Adds a wide `Pop` type (Q96.32, i128-backed, ±~4e28 range) to
`sim_arith` for any quantity that is a count of individuals —
`Cohort` brackets, carrying capacities, food demand, civ
aggregates, migration deltas. The existing `Real` (Q32.32, i64,
±2.1B ceiling) stays for rates, fractions, multipliers, and
everything else; mixed arithmetic ops (`Pop * Real -> Pop`,
`Pop / Pop -> Real`, `Pop / Real -> Pop`) keep the type
boundaries explicit and compiler-checked.

Then retunes the calibration so the new headroom is actually
useful:

  * Baseline `carrying_capacity_per_unit` 2,500 -> 50,000 (20x).
    Earth-equivalent stone-age cell now holds ~50k people.
  * Tier-2 capacity multipliers: BulkCultivation x2 -> x5,
    AnimalSymbiosis x1.5 -> x2.5 (agriculture is the headline
    demographic-transition jump).
  * Tier-4 industrial-era: Mechanisation x3 -> x10,
    ChemicalSynthesis x1.2 -> x3, MedicalIntervention x1.1 -> x2.
  * Tier-5 information-age: AdvancedMedicine x1.2 -> x3,
    GeneticManipulation x1.2 -> x2.
  * Full tech stack now spans ~8,000x (vs ~99x before), matching
    the real paleolithic -> modern density growth of ~1000-5000x.
  * Constructor default `carrying_capacity_per_unit` 500 ->
    50,000 in lockstep.
  * Emergent-civ founding band 20 -> 200 people so a fresh civ
    starts at ~0.4% of cell cap rather than 0.04%.

Wire format: `Pop` raw bits (i128) are serialized as JSON
decimal strings via `pop_bits_serde` / `pop_bits_vec_serde`
helpers because `serde`'s internal `Content` buffer (used by
`#[serde(tag = "kind")]` enums) has no I128 variant. String
encoding round-trips cleanly and is portable across every JSON
consumer.

Trajectory on a typical Earth-equivalent seed:

  | Era         | Tools unlocked | Per-cell cap | Single-civ pop |
  |-------------|----------------|--------------|----------------|
  | Paleolithic | none           | ~50k         | ~1-5M          |
  | Agricultural| tier-2         | ~700k        | ~20-70M        |
  | Industrial  | tier-4         | ~60M         | ~2-6B          |
  | Modern      | tier-5         | ~400M        | ~tens of B+    |

Q96.32 headroom (±~4e28) carries even speculative future-age
megacivilizations without saturation.

All workspace tests pass (`cargo test --workspace --release`).
Empirical 150-year seed-495 smoke run produces sensible
stone-age civs at ~270-800 founding pop; longer runs reach the
projected trajectories above as the tech tree fills.

+++ sim/core/src/phases.rs
@@ -892,23 +892,23 @@ pub(crate) fn population_phase<E: Emitter>(
         }
         nomads::absorb_into_civ(nomad_pops, civ, gained, &species.biology);
         let claimed_sorted = claimed_cells_for_event(civ);
-        let cell_populations_q32: Vec<i64> = claimed_sorted
+        let cell_populations_q32: Vec<i128> = claimed_sorted
             .iter()
             .map(|c| {
                 civ.region_cohorts
                     .get(c)
-                    .map_or(0, |cohort| cohort.total().raw().to_bits())
+                    .map_or(0i128, |cohort| cohort.total().raw().to_bits())
             })
             .collect();
         // Per-cell carrying capacity (cell_capacity in
-        // Real/Q32.32). Mirrors `cell_populations_q32` order so
+        // Pop/Q96.32). Mirrors `cell_populations_q32` order so
         // the viewport renderer can read each cell's pop / cap
         // ratio for the pop-digit scale (digit 9 = saturated,
         // 0 = empty). Same `cell_capacity` formula used by
         // `step_population_per_cell` upstream — tech × terrain ×
         // seasonal × biosphere — so the digit reflects the cap
         // the population step actually evolves toward.
-        let cell_capacities_q32: Vec<i64> = claimed_sorted
+        let cell_capacities_q32: Vec<i128> = claimed_sorted
             .iter()
             .map(|&c| civ.cell_capacity(state, c, tick, planet).raw().to_bits())
             .collect();
@@ -1156,15 +1156,15 @@ pub(crate) fn culture_flip_phase<E: Emitter>(
     let mut events: Vec<CivTerritoryChanged> = Vec::new();
     for civ in civs.iter().filter(|c| touched.contains(&c.id)) {
         let claimed_sorted = claimed_cells_for_event(civ);
-        let cell_pops: Vec<i64> = claimed_sorted
+        let cell_pops: Vec<i128> = claimed_sorted
             .iter()
             .map(|c| {
                 civ.region_cohorts
                     .get(c)
-                    .map_or(0, |cohort| cohort.total().raw().to_bits())
+                    .map_or(0i128, |cohort| cohort.total().raw().to_bits())
             })
             .collect();
-        let cell_caps: Vec<i64> = claimed_sorted
+        let cell_caps: Vec<i128> = claimed_sorted
             .iter()
             .map(|&c| civ.cell_capacity(state, c, tick, planet).raw().to_bits())
             .collect();
